create high noise configuration (#520)
Category: feature
JIRA issue: MIC-5865

Define a function which will return a custom config with a fixed value for all row noising values and for cell probabilities in column noising and default values otherwise.

Testing
Checked that function was returning a config that looked as expected for ACS and census.